strip out non-teddy engines

+++ src/nfa/nfa_rev_api.h
@@ -104,7 +104,7 @@ size_t nfaRevAccel_i(const struct NFA *nfa, const u8 *buffer, size_t length) {
         break;
     case ACCEL_RDEOD:
         DEBUG_PRINTF("ACCEL_RDEOD\n");
-        if (length < nfa->rAccelOffset + 1) {
+        if (length < (size_t)nfa->rAccelOffset + 1) {
             break;
         }
         if (unaligned_load_u16(buffer + length - nfa->rAccelOffset) !=
@@ -114,7 +114,7 @@ size_t nfaRevAccel_i(const struct NFA *nfa, const u8 *buffer, size_t length) {
         break;
     case ACCEL_RDEOD_NOCASE:
         DEBUG_PRINTF("ACCEL_RDEOD_NOCASE\n");
-        if (length < nfa->rAccelOffset + 1) {
+        if (length < (size_t)nfa->rAccelOffset + 1) {
             break;
         }
         if ((unaligned_load_u16(buffer + length - nfa->rAccelOffset) &
